Appian Loan Application
This contains that complete loan application that was developed within Appian. The task was to ensure each personas can complete their task. Applicant: to create a profile with the company, sign up for a loan (create a loan application), and view their loan application status and loans. Loan Officer: to view the loan applications, update the status of the loan application, and send an email once the loan was approved. Account Manager: Only see the approved applications, create a loan that auto populates the information from the loan application onto the loan, and view the reports for the overall company. Bank Executive: To see the data of the entire business, knowing how many active loans there are and other data they may need.